front: Take the hover ring and the row's highlight out of React.
Hovering an object's row still cost more than hovering a place, and the
difference was two things React was carrying that it did not need to.

The map's hover ring was `useState` inside `ExperienceMarkers`, so every
hover re-rendered the component that *is* the map's sources and layers,
and react-map-gl reconciled all of them. It matters more for an object
than for a place because an object rings all of its places — 93 for the
Historic Centre of Saint Petersburg against one. The ring is now written
straight to its source, the way Discover's map has always drawn it.

A row's highlight was a re-render for one background colour, and the
pointer's own case was already pure CSS: React was carrying only the
map→list direction and paying two rows' MUI subtrees per mouse move for
it. The row marks itself with `data-hovered` from its own subscription.

Profiled on Northwestern Federal District, one hover of an object's row:
437 fibers over two commits, then 296 after the ring, then **109 in one**
— the same as hovering a place, and the only thing rendering is the
preview card, which is the thing that changes.

Two more, found with the same profiler while measuring a scroll:

- `CurationDialog` and `AddExperienceDialog` are mounted for as long as
  the list is, closed or not, and re-rendered with it. One wheel scroll
  of Europe's 661 rows spent 25 ms on dialogs nobody had opened. Both are
  memoised now, and the list's close handlers made stable, rather than
  gated on `open` — a gated dialog cannot fade out.
- Closing a card no longer flies the camera back to the whole region.
  A reader who closes a card has usually panned or zoomed for themselves,
  and since #553 refitting did more than move the camera: republishing
  the view changes which rows are listed, so closing one card quietly
  re-filtered the list. That was the last use of `shouldFitRegion`, so
  the trigger, its context state and the effect that consumed it are gone.
