+++ docs/tech/maplibre-patterns.md
@@ -50,7 +50,11 @@ map.off('click', onMarkerClick);   // cleanup takes the same shape
 
 ### The rule
 
-**One handler, one registration.** Registering the same handler across layers that can overlap is what runs it twice — a single-layer registration is one listener whatever the handler does, which is why `map.on('click', 'clusters', …)` navigating on its own layer is fine. Where layers do overlap, only idempotent handlers may stay per layer: `mousemove` and `mouseleave` may, because painting the same ring twice is painting it once. Anything that toggles, counts, navigates or mutates must be registered once, on the map.
+**One handler, one registration.** Registering the same handler across layers that can overlap is what runs it twice — a single-layer registration is one listener whatever the handler does, which is why `map.on('click', 'clusters', …)` navigating on its own layer is fine. Where layers do overlap, only idempotent handlers may stay per layer: `mousemove` may, because painting the same ring twice is painting it once. Anything that toggles, counts, navigates or mutates must be registered once, on the map.
+
+`mouseleave` is not one of them, and idempotency is not what makes it safe. Over overlapping layers it fires a *spurious* leave: moving from a pin onto its own count badge leaves the marker layer while the pointer has not left the pin, and clearing there — once is already too many — takes the popup and the hover ring off something the reader is still pointing at. A per-layer `mouseleave` over layers that overlap needs a condition of its own: that the thing under the point is not the thing the hover is already on. That means one leave handler across every overlapping layer, reading one key space — not one per layer, each blind to the others. `useMarkerInteractions.ts` carries that check for its four: a leave arriving *after* a sibling layer's `mousemove` has set the hover would otherwise clear what the reader has just arrived at, and within one DOM move the handlers run in registration order.
+
+Test it against both paths. MapLibre's delegated `mouseleave` fires from a `mousemove` that no longer hits the layer *and* from a `mouseout` when the pointer leaves the canvas — and only the first guarantees the point is off the feature. A `mouseout` carries the point the pointer left *from*, which is still on the pin whenever it left onto an overlay drawn above the map, and after it `mousein` is false, so no further leave arrives until a feature is entered again. A same-target check that trusts that point leaves the hover lit for good; check `e.originalEvent.type` before trusting it.
 
 A global registration gives up one thing the delegated form did for free: MapLibre checks `getLayer(layerId)` before querying, so a click while those layers are absent did nothing. Naming a missing layer in `queryRenderedFeatures` fires an `ErrorEvent` and logs — and layers do go absent, since a component that returns `null` while its data loads unmounts its `<Source>`/`<Layer>` tree while the handler effect stays attached. Open the handler with the same check:
