+++ docs/vision/vision.md
@@ -16,7 +16,7 @@ A visitor can:
 
 - **Browse the map** — see the world divided into regions (countries, sub-regions, custom groupings) rendered as interactive vector tiles
 - **Explore experiences** — browse UNESCO World Heritage Sites, top museums, public art & monuments organized by region through the Discover page
-- **See what exists** — understand the breadth of experiences available in any region, with counts, categories, images, and descriptions. A region is shown whole rather than as a first page of it: the map draws every experience assigned there, and the list reaches all of them — it answers about what the map is showing, with one line offering the rest of the region back. Neither is a first page, so a category is never represented only by its alphabetically-early members. Whole no longer means slow: a region holding hundreds of places used to lock the tab for a couple of seconds while it drew every row at once, and now draws the rows you are looking at as you scroll — the same list, the same scrollbar covering all of it, without the freeze. Opening a place's card no longer makes the list shuffle underneath you either. A card is not shown until everything in it is ready, so it opens once, at the size it keeps, and the rest of the list stays where you left it — where before it grew in steps and then jumped again when the picture arrived. While a card is being fetched the row you clicked says so quietly, without moving, and if your pointer rested on that row for a moment beforehand its picture is already waiting, which is the slowest part of the wait by far. An open card of a sprawling site now shows the first twenty of its places with a line offering the rest — the Historic Centre of Saint Petersburg has 112, and listing all of them at once was half a second of work before the card could appear. Running your pointer down that list is answered at once, on the row and on the map alike: it used to lag by three to five rows and light the map a second or two later, which made a card of many places feel stuck rather than slow
+- **See what exists** — understand the breadth of experiences available in any region, with counts, categories, images, and descriptions. A region is shown whole rather than as a first page of it: the map draws every experience assigned there, and the list reaches all of them — it answers about what the map is showing, with one line offering the rest of the region back. Neither is a first page, so a category is never represented only by its alphabetically-early members. Whole no longer means slow: a region holding hundreds of places used to lock the tab for a couple of seconds while it drew every row at once, and now draws the rows you are looking at as you scroll — the same list, the same scrollbar covering all of it, without the freeze. Opening a place's card no longer makes the list shuffle underneath you either. A card is not shown until everything in it is ready, so it opens once, at the size it keeps, and the rest of the list stays where you left it — where before it grew in steps and then jumped again when the picture arrived. While a card is being fetched the row you clicked says so quietly, without moving, and if your pointer rested on that row for a moment beforehand its picture is already waiting, which is the slowest part of the wait by far. An open card of a sprawling site now shows the first twenty of its places with a line offering the rest — the Historic Centre of Saint Petersburg has 112, and listing all of them at once was half a second of work before the card could appear. Running your pointer down that list is answered at once, on the row and on the map alike: it used to lag by three to five rows and light the map a second or two later, which made a card of many places feel stuck rather than slow. Closing a card leaves the map where you put it, too — it used to fly back out to the whole region, throwing away the view you had zoomed into and, with it, changing which rows the list was showing you
 - **Switch world views** — see the world organized in different ways: any custom regional grouping an admin has published. Administrative geography is reachable this way only if an admin publishes a world view built from it (see the base layer import below); the built-in administrative hierarchy itself stays admin-only. The address bar follows the choice and the choice follows the address bar: a link carrying a world view opens on it, and pasting or editing one in a tab that is already open switches to it
 - **What just arrived** — a **New** mark on things you could first see recently, not on things that happen to be recent rows in a database. It stays for a while — how long depends on the source, since some publish monthly and some rarely — and then it fades on its own, per thing. It is not tied to a batch: a source that updates weekly shows several weeks' worth of new things at a time rather than only its latest run's, and a thing held back for review carries its mark from the day you could see it rather than from the day an importer found it
 - **One answer to where something is** — a place you can actually go to. Some sites are one thing in a list and another on the map: the world's official register publishes a point for the *property*, which for a rainforest spread across a state can be 191 km from any part of it you could visit. What you are shown now is never that locator: it is a real place of the site. A site made of many parts is drawn as its parts, on the map and in Discover alike — the rainforest is its forty forests, the rock art is the shelters you can walk to — so what the map says is here is what is here. Where one point is unavoidable, it is the real place nearest to what the source published
